refactor: [FE] SketchbookCard 컴포넌트 리팩터링 2
- Intent
SketchbookCard를 결과 페이지에서 재사용하고 싶어서 우선 기존 로직들을 훅으로 분리하여 리팩터링 했어요.
