+++ frontend/src/components/RoundNumberAndTimer.tsx
@@ -1,0 +1,13 @@
+import RoundNumber from '@components/RoundNumber';
+import Timer from '@components/Timer';
+
+function RoundNumberAndTimer() {
+    return (
+        <>
+            <RoundNumber />
+            <Timer />
+        </>
+    );
+}
+
+export default RoundNumberAndTimer;
