+++ frontend/src/components/SketchbookCard.tsx
@@ -1,36 +1,25 @@
+import { ReactNode } from 'react';
 import styled from 'styled-components';
 import { Center } from '@styles/styled';
-import { ReactComponent as Sketchbook } from '@assets/sketchbook.svg';
-import { useRecoilValue } from 'recoil';
-import { isQuizTypeDrawState, roundNumberState, quizSubmitState } from '@atoms/game';
-import useCanvas from '@hooks/useCanvas';
+import { ReactComponent as SketchbookImg } from '@assets/sketchbook.svg';
 import Card from '@components/Card';
-import Timer from '@components/Timer';
-import DrawingTools from '@components/DrawingTools';
-import usePrevQuizReply from '@hooks/usePrevQuizReply';
 
-function SketchbookCard() {
-    const { canvasRef, ...restProps } = useCanvas();
-    const isDraw = useRecoilValue(isQuizTypeDrawState);
-    const { curRound, maxRound } = useRecoilValue(roundNumberState);
-    const quizSubmitted = useRecoilValue(quizSubmitState);
-    const { renderPrevUserQuizReply } = usePrevQuizReply();
+interface SketchbookCardType {
+    left?: ReactNode;
+    center: ReactNode;
+    right?: ReactNode;
+}
 
+function SketchbookCard({ left, center, right }: SketchbookCardType) {
     return (
         <Card>
             <Container>
-                <GameStateSection>
-                    <GameTurn>
-                        {curRound}/{maxRound}
-                    </GameTurn>
-                    <Timer />
-                </GameStateSection>
+                <LeftSide>{left}</LeftSide>
                 <SketchbookWrapper>
-                    <Sketchbook />
-                    <Canvas ref={canvasRef} />
-                    {renderPrevUserQuizReply()}
+                    <SketchbookImg />
+                    {center}
                 </SketchbookWrapper>
-                {isDraw && !quizSubmitted ? <DrawingTools restProps={restProps} /> : <div />}
+                <RightSide>{right}</RightSide>
             </Container>
         </Card>
     );
@@ -40,32 +29,11 @@ export default SketchbookCard;
 
 const Container = styled(Center)`
     padding: 44px 38px 0 28px;
-
-    > div:last-of-type {
-        width: 100%;
-    }
 `;
 
-const GameStateSection = styled.div`
+const LeftSide = styled.div`
     display: flex;
     align-items: end;
-`;
-
-const GameTurn = styled.div`
-    height: 45px;
-    margin-right: 32px;
-    background: ${({ theme }) => theme.gradation.primaryLightBrown};
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-    text-fill-color: transparent;
-    -webkit-text-stroke: 1px ${({ theme }) => theme.color.blackT1};
-    text-stroke: 1px ${({ theme }) => theme.color.blackT1};
-    font-family: 'Sniglet', cursive;
-    font-weight: 800;
-    font-size: 1.75rem;
-    letter-spacing: 0.05rem;
-    transform: translateY(16px);
 `;
 
 const SketchbookWrapper = styled.div`
@@ -73,6 +41,6 @@ const SketchbookWrapper = styled.div`
     margin: 0 30px;
 `;
 
-const Canvas = styled.canvas`
-    ${({ theme }) => theme.layout.sketchBook};
+const RightSide = styled.div`
+    width: 100%;
 `;
