+++ frontend/src/components/Timer.tsx
@@ -51,6 +51,7 @@ export default Timer;
 
 const Container = styled(Center)`
     flex-direction: column;
+    margin-left: 32px;
 `;
 
 const ProgressBar = styled.div`
