+++ frontend/src/pages/Game.tsx
@@ -1,20 +1,28 @@
+import { useEffect } from 'react';
 import styled from 'styled-components';
 import { ScaledDiv, ScaledSection } from '@styles/styled';
+import { onCompleteGame, onCountSubmittedQuiz } from '@game/NetworkServiceUtils';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { gameResultState, isQuizTypeDrawState, quizSubmitState } from '@atoms/game';
+import QuizReplySection from '@components/QuizReplySection';
+import DrawingTools from '@components/DrawingTools';
+import RoundNumberAndTimer from '@components/RoundNumberAndTimer';
 import SketchbookCard from '@components/SketchbookCard';
 import GameUsers from '@components/GameUsers';
 import MicButton from '@components/MicButton';
 import CameraButton from '@components/CameraButton';
-import SmallLogo from '@assets/logo-s.png';
 import useMovePage from '@hooks/useMovePage';
-import QuizReplySection from '@components/QuizReplySection';
-import { useEffect } from 'react';
-import { onCompleteGame, onCountSubmittedQuiz } from '@game/NetworkServiceUtils';
-import { useSetRecoilState } from 'recoil';
-import { gameResultState } from '@atoms/game';
+import SmallLogo from '@assets/logo-s.png';
+import usePrevQuizReply from '@hooks/usePrevQuizReply';
+import useCanvas from '@hooks/useCanvas';
 
 function Game() {
     const [setPage] = useMovePage();
     const setGameResult = useSetRecoilState(gameResultState);
+    const { canvasRef, ...restProps } = useCanvas();
+    const { renderPrevUserQuizReply } = usePrevQuizReply();
+    const isDraw = useRecoilValue(isQuizTypeDrawState);
+    const quizSubmitted = useRecoilValue(quizSubmitState);
 
     useEffect(() => {
         onCountSubmittedQuiz();
@@ -26,7 +34,16 @@ function Game() {
             <Container>
                 <GameUsers />
                 <SketchbookSection>
-                    <SketchbookCard />
+                    <SketchbookCard
+                        left={<RoundNumberAndTimer />}
+                        center={
+                            <>
+                                <Canvas ref={canvasRef} />
+                                {renderPrevUserQuizReply()}
+                            </>
+                        }
+                        right={isDraw && !quizSubmitted && <DrawingTools restProps={restProps} />}
+                    />
                     <QuizReplySection />
                 </SketchbookSection>
             </Container>
@@ -68,4 +85,8 @@ const LogoWrapper = styled(ScaledDiv)`
     position: absolute;
     bottom: 40px;
     cursor: pointer;
+`;
+
+const Canvas = styled.canvas`
+    ${({ theme }) => theme.layout.sketchBook};
 `;
